V2.0.0

+++ projector/tool/saveDocumentVersion.php
@@ -17,53 +17,60 @@ header ('Content-Type: text/html; charset=UTF-8');
 $error=false;
 $documentVersionLink="";
 $uploadedFile=false;
-if (array_key_exists('documentVersionLink',$_REQUEST)) {
-	$documentVersionLink=$_REQUEST['documentVersionLink'];
-}
-if (array_key_exists('documentVersionFile',$_FILES)) {
-  $uploadedFile=$_FILES['documentVersionFile'];
-} else if ($documentVersionLink!='') {
-	// OK Link instead of file
-} else {
-  echo htmlGetErrorMessage(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
-  errorLog(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
-  $error=true; 
-}
-if ($uploadedFile and $documentVersionLink and $uploadedFile['name']) {
-	echo htmlGetWarningMessage(i18n('errorFileOrLink',null));
-	$error=true; 
-}
-if (! $error and $uploadedFile and $uploadedFile['name']) {
-  if ( $uploadedFile['error']!=0 ) {
-    switch ($uploadedFile['error']) {
-      case 1:
-        echo htmlGetErrorMessage(i18n('errorTooBigFile',array(ini_get('upload_max_filesize'),'upload_max_filesize')));
-        errorLog(i18n('errorTooBigFile',array(ini_get('upload_max_filesize'),'upload_max_filesize')));
-        break; 
-      case 2:
-        echo htmlGetErrorMessage(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
-        errorLog(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
-        break;  
-      case 4:
-        echo htmlGetWarningMessage(i18n('errorNoFile'));
-        errorLog(i18n('errorNoFile'));
-        break;  
-      default:
-        echo htmlGetErrorMessage(i18n('errorUploadFile',array($uploadedFile['error'])));
-        errorLog(i18n('errorUploadFile',array($uploadedFile['error'])));
-        break;
-    }
-    $error=true; 
-  }
-}
-if (! $error and $uploadedFile and !$documentVersionLink) {
-  if (! $uploadedFile['name']) {
-    echo htmlGetWarningMessage(i18n('errorNoFile'));
-    errorLog(i18n('errorNoFile'));
-    $error=true; 
-  }
+
+$documentVersionId=null;
+if (array_key_exists('documentVersionId',$_REQUEST)) {
+    $documentVersionId=$_REQUEST['documentVersionId'];
 }
 
+if (! $documentVersionId) { // Get file only on insert
+	if (array_key_exists('documentVersionLink',$_REQUEST)) {
+		$documentVersionLink=$_REQUEST['documentVersionLink'];
+	}
+	if (array_key_exists('documentVersionFile',$_FILES)) {
+	  $uploadedFile=$_FILES['documentVersionFile'];
+	} else if ($documentVersionLink!='') {
+		// OK Link instead of file
+	} else {
+	  echo htmlGetErrorMessage(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
+	  errorLog(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
+	  $error=true; 
+	}
+	if ($uploadedFile and $documentVersionLink and $uploadedFile['name']) {
+		echo htmlGetWarningMessage(i18n('errorFileOrLink',null));
+		$error=true; 
+	}
+	if (! $error and $uploadedFile and $uploadedFile['name']) {
+	  if ( $uploadedFile['error']!=0 ) {
+	    switch ($uploadedFile['error']) {
+	      case 1:
+	        echo htmlGetErrorMessage(i18n('errorTooBigFile',array(ini_get('upload_max_filesize'),'upload_max_filesize')));
+	        errorLog(i18n('errorTooBigFile',array(ini_get('upload_max_filesize'),'upload_max_filesize')));
+	        break; 
+	      case 2:
+	        echo htmlGetErrorMessage(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
+	        errorLog(i18n('errorTooBigFile',array($paramAttachementMaxSize,'$paramAttachementMaxSize')));
+	        break;  
+	      case 4:
+	        echo htmlGetWarningMessage(i18n('errorNoFile'));
+	        errorLog(i18n('errorNoFile'));
+	        break;  
+	      default:
+	        echo htmlGetErrorMessage(i18n('errorUploadFile',array($uploadedFile['error'])));
+	        errorLog(i18n('errorUploadFile',array($uploadedFile['error'])));
+	        break;
+	    }
+	    $error=true; 
+	  }
+	}
+	if (! $error and $uploadedFile and !$documentVersionLink) {
+	  if (! $uploadedFile['name']) {
+	    echo htmlGetWarningMessage(i18n('errorNoFile'));
+	    errorLog(i18n('errorNoFile'));
+	    $error=true; 
+	  }
+	}
+}
 $documentVersionNewVersion=null;
 if (! array_key_exists('documentVersionNewVersion',$_REQUEST)) {
     echo htmlGetErrorMessage('documentVersionNewVersion parameter not found in REQUEST');
@@ -92,6 +99,13 @@ if (! array_key_exists('documentId',$_REQUEST)) {
 }
 $documentId=$_REQUEST['documentId'];
 
+$documentVersionDate=null;
+if (! array_key_exists('documentVersionDate',$_REQUEST)) {
+    echo htmlGetErrorMessage('documentVersionDate parameter not found in REQUEST');
+    $error=true;
+}
+$documentVersionDate=$_REQUEST['documentVersionDate'];
+
 $documentVersionDescription=null;
 if (! array_key_exists('documentVersionDescription',$_REQUEST)) {
     echo htmlGetErrorMessage('documentVersionDescription parameter not found in REQUEST');
@@ -106,26 +120,19 @@ if (! array_key_exists('documentVersionNewVersionDisplay',$_REQUEST)) {
 }
 $documentVersionNewVersionDisplay=$_REQUEST['documentVersionNewVersionDisplay'];
 
-
-$documentVersionId=null;
-if (array_key_exists('documentVersionId',$_REQUEST)) {
-    $documentVersionId=$_REQUEST['documentVersionId'];
-}
-
 if (! $error) {
-  $doc=New Document($documentId);
-  $dir=New DocumentDirectory($doc->idDocumentDirectory);
-	
 	$dv=new DocumentVersion($documentVersionId);
   $dv->idDocument=$documentId;
   $dv->idAuthor=$user->id;
-  $dv->versionDate=date("Y-m-d");
-  if ($documentVersionLink) {
-  	
-  } else {
-    $dv->fileName=basename($uploadedFile['name']);
-    $dv->mimeType=$uploadedFile['type'];
-    $dv->fileSize=$uploadedFile['size'];
+  $dv->versionDate=$documentVersionDate;
+  if (! $documentVersionId) {
+	  if ($documentVersionLink) {
+	  	
+	  } else {
+	    $dv->fileName=basename($uploadedFile['name']);
+	    $dv->mimeType=$uploadedFile['type'];
+	    $dv->fileSize=$uploadedFile['size'];
+	  }
   }
   $dv->description=$documentVersionDescription;
   $dv->version=$documentVersionNewVersion;
@@ -134,39 +141,30 @@ if (! $error) {
   $dv->name=$documentVersionNewVersionDisplay;
   $result=$dv->save();
   $newId= $dv->id;
-  
-  if ( ($doc->version==null) 
-    or ( $dv->version>$doc->version ) 
-    or ( $dv->version==$doc->version and $dv->revision>$doc->revision) 
-    or ( $dv->version==$doc->version and $dv->revision==$doc->revision and $dv->draft>$doc->draft) ) {
-    	$doc->version=$dv->version;
-    	$doc->revision=$dv->revision;
-    	$doc->draft=$dv->draft;
-    	$doc->save();
-    }
 }
 
-if (! $error and !$documentVersionLink) {
-  $root=Parameter::getGlobalParameter('documentRoot');
-	$uploaddir = $root . $dir->location ;
-	$split=explode('/',$uploaddir);
-	$dir='';
-	foreach ($split as $dirElt) { 
-		$dir.=$dirElt.'/';
-    if (! file_exists($dir)) {
-      mkdir($dir);
-    }
+if (! $documentVersionId) {
+	if (! $error and !$documentVersionLink ) {
+		$uploadfile = $dv->getUploadFileName();
+		$split=explode($paramPathSeparator,$uploadfile);
+		unset($split[count($split)-1]);
+		$dir='';
+		foreach ($split as $dirElt) { 
+			$dir.=$dirElt.'/';
+	    if (! file_exists($dir)) {
+	      mkdir($dir);
+	    }
+		}
+	  if ( ! move_uploaded_file($uploadedFile['tmp_name'], $uploadfile)) {
+	     echo htmlGetErrorMessage(i18n('errorUploadFile','hacking ?'));
+	     errorLog(i18n('errorUploadFile','hacking ?'));
+	     $error=true;
+	     $dv->delete(); 
+	  } else {
+	    //$dv->subDirectory=$uploaddir;
+	    //$otherResult=$dv->save();
+	  }
 	}
-  $uploadfile = $uploaddir . $paramPathSeparator . basename($uploadedFile['name'] . '.' . $dv->id);
-  if ( ! move_uploaded_file($uploadedFile['tmp_name'], $uploadfile)) {
-     echo htmlGetErrorMessage(i18n('errorUploadFile','hacking ?'));
-     errorLog(i18n('errorUploadFile','hacking ?'));
-     $error=true;
-     $dv->delete(); 
-  } else {
-    //$dv->subDirectory=$uploaddir;
-    //$otherResult=$dv->save();
-  }
 }
 if (! $error) {
   // Message of correct saving
